+++ src/main/java/ch/ethz/matsim/av/dispatcher/multi_od_heuristic/SerialAggregateRideAppender.java
@@ -139,7 +139,7 @@ public class SerialAggregateRideAppender implements AggregateRideAppender {
                 AVPickupTask pickupTask = new AVPickupTask(
                         currentTime,
                         currentTime + timing.getPickupDurationPerStop(),
-                        pickup.getFromLink(),
+                        pickup.getFromLink(), Double.NEGATIVE_INFINITY,
                         Arrays.asList(pickup)
                 );